Relocate classes and fix package name
