+++ kafka-connect/kafka-connect/src/main/java/org/apache/iceberg/connect/IcebergSinkTask.java
@@ -18,10 +18,13 @@
  */
 package org.apache.iceberg.connect;
 
+import io.tabular.iceberg.connect.IcebergSinkConfig;
+import io.tabular.iceberg.connect.channel.Coordinator;
+import io.tabular.iceberg.connect.channel.Worker;
+import io.tabular.iceberg.connect.data.Utilities;
 import java.util.Collection;
 import java.util.Map;
 import org.apache.iceberg.catalog.Catalog;
-import org.apache.iceberg.relocated.com.google.common.base.Preconditions;
 import org.apache.iceberg.relocated.com.google.common.collect.ImmutableMap;
 import org.apache.kafka.clients.consumer.OffsetAndMetadata;
 import org.apache.kafka.common.TopicPartition;
@@ -35,12 +38,12 @@ public class IcebergSinkTask extends SinkTask {
   private static final Logger LOG = LoggerFactory.getLogger(IcebergSinkTask.class);
 
   private IcebergSinkConfig config;
-  private Catalog catalog;
-  private Committer committer;
+  private Coordinator coordinator;
+  private Worker worker;
 
   @Override
   public String version() {
-    return IcebergSinkConfig.version();
+    return IcebergSinkConfig.getVersion();
   }
 
   @Override
@@ -50,60 +53,78 @@ public class IcebergSinkTask extends SinkTask {
 
   @Override
   public void open(Collection<TopicPartition> partitions) {
-    Preconditions.checkArgument(catalog == null, "Catalog already open");
-    Preconditions.checkArgument(committer == null, "Committer already open");
+    Catalog catalog = Utilities.loadCatalog(config);
 
-    catalog = CatalogUtils.loadCatalog(config);
-    committer = CommitterFactory.createCommitter(config);
-    committer.start(catalog, config, context);
+    if (isLeader(partitions)) {
+      LOG.info("Task elected leader, starting commit coordinator");
+      coordinator = new Coordinator(catalog, config);
+      coordinator.start();
+    }
+    LOG.info("Starting commit worker");
+    worker = new Worker(catalog, config, context);
+    worker.syncCommitOffsets();
+    worker.start();
+  }
+
+  private boolean isLeader(Collection<TopicPartition> partitions) {
+    // there should only be one worker assigned partition 0 of the first
+    // topic, so elect that one the leader
+    String firstTopic = config.getTopics().first();
+    return partitions.stream()
+        .filter(tp -> tp.topic().equals(firstTopic))
+        .anyMatch(tp -> tp.partition() == 0);
   }
 
   @Override
   public void close(Collection<TopicPartition> partitions) {
-    close();
-  }
-
-  private void close() {
-    if (committer != null) {
-      committer.stop();
-      committer = null;
+    if (worker != null) {
+      worker.stop();
+      worker = null;
     }
-
-    if (catalog != null) {
-      if (catalog instanceof AutoCloseable) {
-        try {
-          ((AutoCloseable) catalog).close();
-        } catch (Exception e) {
-          LOG.warn("An error occurred closing catalog instance, ignoring...", e);
-        }
-      }
-      catalog = null;
+    if (coordinator != null) {
+      coordinator.stop();
+      coordinator = null;
     }
   }
 
   @Override
   public void put(Collection<SinkRecord> sinkRecords) {
-    if (committer != null) {
-      committer.save(sinkRecords);
+    if (sinkRecords != null && !sinkRecords.isEmpty() && worker != null) {
+      worker.save(sinkRecords);
     }
+    coordinate();
   }
 
   @Override
   public void flush(Map<TopicPartition, OffsetAndMetadata> currentOffsets) {
-    if (committer != null) {
-      committer.save(null);
-    }
+    coordinate();
   }
 
   @Override
   public Map<TopicPartition, OffsetAndMetadata> preCommit(
       Map<TopicPartition, OffsetAndMetadata> currentOffsets) {
-    // offset commit is handled by the worker
-    return ImmutableMap.of();
+    if (worker == null) {
+      return ImmutableMap.of();
+    }
+    return worker.getCommitOffsets();
+  }
+
+  private void coordinate() {
+    if (worker != null) {
+      worker.process();
+    }
+    if (coordinator != null) {
+      coordinator.process();
+    }
   }
 
   @Override
   public void stop() {
-    close();
+    if (worker != null) {
+      worker.stop();
+    }
+    if (coordinator != null) {
+      coordinator.stop();
+    }
   }
 }
