+++ kafka-connect/kafka-connect/src/main/java/org/apache/iceberg/connect/channel/events/TableName.java
@@ -16,40 +16,62 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package io.tabular.iceberg.connect.channel.events;
+package org.apache.iceberg.connect.channel.events;
 
-import static org.apache.iceberg.avro.AvroSchemaUtil.FIELD_ID_PROP;
-
-import java.util.UUID;
+import io.tabular.iceberg.connect.channel.events.Element;
+import java.util.Arrays;
+import java.util.List;
 import org.apache.avro.Schema;
 import org.apache.avro.SchemaBuilder;
+import org.apache.avro.util.Utf8;
+import org.apache.iceberg.catalog.Namespace;
+import org.apache.iceberg.catalog.TableIdentifier;
 
-public class CommitRequestPayload implements Payload {
+import static java.util.stream.Collectors.toList;
+import static org.apache.iceberg.avro.AvroSchemaUtil.FIELD_ID_PROP;
 
-  private UUID commitId;
+public class TableName implements Element {
+
+  private List<String> namespace;
+  private String name;
   private Schema avroSchema;
 
   public static final Schema AVRO_SCHEMA =
       SchemaBuilder.builder()
-          .record(CommitRequestPayload.class.getName())
+          .record(TableName.class.getName())
           .fields()
-          .name("commitId")
+          .name("namespace")
           .prop(FIELD_ID_PROP, DUMMY_FIELD_ID)
-          .type(UUID_SCHEMA)
+          .type()
+          .array()
+          .items()
+          .stringType()
+          .noDefault()
+          .name("name")
+          .prop(FIELD_ID_PROP, DUMMY_FIELD_ID)
+          .type()
+          .stringType()
           .noDefault()
           .endRecord();
 
-  public CommitRequestPayload(Schema avroSchema) {
+  public static TableName of(TableIdentifier tableIdentifier) {
+    return new TableName(
+        Arrays.asList(tableIdentifier.namespace().levels()), tableIdentifier.name());
+  }
+
+  public TableName(Schema avroSchema) {
     this.avroSchema = avroSchema;
   }
 
-  public CommitRequestPayload(UUID commitId) {
-    this.commitId = commitId;
+  public TableName(List<String> namespace, String name) {
+    this.namespace = namespace;
+    this.name = name;
     this.avroSchema = AVRO_SCHEMA;
   }
 
-  public UUID getCommitId() {
-    return commitId;
+  public TableIdentifier toIdentifier() {
+    Namespace icebergNamespace = Namespace.of(namespace.toArray(new String[0]));
+    return TableIdentifier.of(icebergNamespace, name);
   }
 
   @Override
@@ -58,10 +80,15 @@ public class CommitRequestPayload implements Payload {
   }
 
   @Override
+  @SuppressWarnings("unchecked")
   public void put(int i, Object v) {
     switch (i) {
       case 0:
-        this.commitId = (UUID) v;
+        this.namespace =
+            v == null ? null : ((List<Utf8>) v).stream().map(Utf8::toString).collect(toList());
+        return;
+      case 1:
+        this.name = v == null ? null : v.toString();
         return;
       default:
         // ignore the object, it must be from a newer version of the format
@@ -72,7 +99,9 @@ public class CommitRequestPayload implements Payload {
   public Object get(int i) {
     switch (i) {
       case 0:
-        return commitId;
+        return namespace;
+      case 1:
+        return name;
       default:
         throw new UnsupportedOperationException("Unknown field ordinal: " + i);
     }
