+++ kafka-connect/kafka-connect/src/main/java/org/apache/iceberg/connect/channel/events/TopicPartitionOffset.java
@@ -16,12 +16,13 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package io.tabular.iceberg.connect.channel.events;
+package org.apache.iceberg.connect.channel.events;
 
-import static org.apache.iceberg.avro.AvroSchemaUtil.FIELD_ID_PROP;
-
+import io.tabular.iceberg.connect.channel.events.Element;
 import org.apache.avro.Schema;
 import org.apache.avro.SchemaBuilder;
+
+import static org.apache.iceberg.avro.AvroSchemaUtil.FIELD_ID_PROP;
 
 public class TopicPartitionOffset implements Element {
