+++ kafka-connect/kafka-connect/src/main/java/org/apache/iceberg/connect/data/RecordWrapper.java
@@ -16,8 +16,9 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package io.tabular.iceberg.connect.data;
+package org.apache.iceberg.connect.data;
 
+import io.tabular.iceberg.connect.data.Operation;
 import java.util.Map;
 import org.apache.iceberg.data.Record;
 import org.apache.iceberg.types.Types.StructType;
