+++ kafka-connect/kafka-connect/src/main/java/org/apache/iceberg/connect/data/WriterResult.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package io.tabular.iceberg.connect.data;
+package org.apache.iceberg.connect.data;
 
 import java.util.List;
 import java.util.Map;
